prql-python/requirements.txt
pytest==7.0.0
pytest-xdist==2.5.0
pytest-clarity==1.0.1
PyYAML==6.0
